Refactor findByFieldOnlyOne method in Auth model
This commit refactors the `findByFieldOnlyOne` method in the `Auth` model. It updates the SQL query to include a join with the `usuarios_roles` table and retrieves the `nombre` and `rol_id` from the `roles` table. The result is then assigned to the `rol` variable. The `usuario_id` and `rol_id` properties of the `result` object are updated accordingly.